requirements.txt
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
selenium>=4.15.0
python-dotenv>=1.0.0
